Change router Action in Results view
Results view was not configured to send
a player to the new Match Search view
that starts the entire game flow process.
Button was originally configured to send
the user to the Gameplay component that
housed the old sockets middleware and
started the whole quiz flow. Also change
the placeholder text in Gameplay when a
quiz question hasn't been received from
the server.